Nits and bits
Change
------

Setup tarpaulin and added a few unit-tests for better coverage.

Unfortunately, tarpauling doesn't respect the cargo-target directory
for doctests, which makes them impossibly long to compile in my setup
and slightly taint the output.

Still, manually correcting for the lack of doctests, it seems that the
library is (finally) quite adequately tested.